+++ launcher/StaticComponents.h
@@ -25,40 +25,40 @@
 #ifdef SHAPE_STATIC_LIBS
 
 extern "C" {
-  const shape::ComponentMeta& get_component_shape__TraceFormatService(unsigned long* compiler, unsigned long* typehash);
-  const shape::ComponentMeta& get_component_shape__TraceFileService(unsigned long* compiler, unsigned long* typehash);
-  const shape::ComponentMeta& get_component_shape__LauncherService(unsigned long* compiler, unsigned long* typehash);
-  const shape::ComponentMeta& get_component_shape__ConfigurationService(unsigned long* compiler, unsigned long* typehash);
+  const shape::ComponentMeta& get_component_shape__TraceFormatService(unsigned long* compiler, size_t* typehash);
+  const shape::ComponentMeta& get_component_shape__TraceFileService(unsigned long* compiler, size_t* typehash);
+  const shape::ComponentMeta& get_component_shape__LauncherService(unsigned long* compiler, size_t* typehash);
+  const shape::ComponentMeta& get_component_shape__ConfigurationService(unsigned long* compiler, size_t* typehash);
 }
 
 void staticInit()
 {
   void* component = nullptr;
   unsigned long compiler = 0;
-  unsigned long hashcode = 0;
+  size_t hashcode = 0;
   unsigned long expectedCompiler = (unsigned long)SHAPE_PREDEF_COMPILER;
-  unsigned long expectedHashcode = std::type_index(typeid(shape::ComponentMeta)).hash_code();
+  size_t expectedHashcode = std::type_index(typeid(shape::ComponentMeta)).hash_code();
 
   // get components
   shape::ComponentMap::get().setComponent(&get_component_shape__TraceFormatService(&compiler, &hashcode));
   shape::ComponentMap::get().setComponent(&get_component_shape__TraceFileService(&compiler, &hashcode));
   shape::ComponentMap::get().setComponent(&get_component_shape__LauncherService(&compiler, &hashcode));
-  const shape::ComponentMeta& get_component_shape__ConfigurationService(unsigned long* compiler, unsigned long* typehash);
+  const shape::ComponentMeta& get_component_shape__ConfigurationService(unsigned long* compiler, size_t* typehash);
 }
 
 #else
 extern "C" {
-  const shape::ComponentMeta& get_component_shape__LauncherService(unsigned long* compiler, unsigned long* typehash);
-  const shape::ComponentMeta& get_component_shape__ConfigurationService(unsigned long* compiler, unsigned long* typehash);
+  const shape::ComponentMeta& get_component_shape__LauncherService(unsigned long* compiler, size_t* typehash);
+  const shape::ComponentMeta& get_component_shape__ConfigurationService(unsigned long* compiler, size_t* typehash);
 }
 
 void staticInit()
 {
   void* component = nullptr;
   unsigned long compiler = 0;
-  unsigned long hashcode = 0;
+  size_t hashcode = 0;
   unsigned long expectedCompiler = (unsigned long)SHAPE_PREDEF_COMPILER;
-  unsigned long expectedHashcode = std::type_index(typeid(shape::ComponentMeta)).hash_code();
+  size_t expectedHashcode = std::type_index(typeid(shape::ComponentMeta)).hash_code();
 
   shape::ComponentMap::get().setComponent(&get_component_shape__LauncherService(&compiler, &hashcode));
   shape::ComponentMap::get().setComponent(&get_component_shape__ConfigurationService(&compiler, &hashcode));
